Pop up added when deleting an organization

+++ src/app/pages/organization-list/organization-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {Organization} from '../../models/Organization';
 import {OrganizationService} from '../../services/organization.service';
 import {BaseComponent} from '../../components/base/base.component';
+import {DialogService} from '../../services/dialog.service';
 
 @Component({
   selector: 'app-organization-list',
@@ -13,7 +14,8 @@ export class OrganizationListComponent extends BaseComponent implements OnInit {
   title = 'Organizaciones';
   organizations: Array<Organization> = [];
 
-  constructor(private organizationService: OrganizationService) {
+  constructor(private organizationService: OrganizationService,
+              private dialogService: DialogService) {
     super();
   }
 
@@ -25,12 +27,16 @@ export class OrganizationListComponent extends BaseComponent implements OnInit {
   }
 
   deleteOrganization(organization: Organization) {
-    // TODO add pop up asking if you are sure
-    this.organizationService.deleteOrganization(organization.id.toString())
-      .subscribe(data => {
-        this.setSuccess(`La organización "${organization.name}" fue eliminada`);
-        this.organizations = this.organizations.filter(org => org.id !== organization.id);
-      },
-        error => this.setError('No se pudo eliminar la organización'));
+    this.dialogService.openConfirmDialog('¿Seguro que desea borrar la organización?')
+      .afterClosed().subscribe(response => {
+        if (response) {
+          this.organizationService.deleteOrganization(organization.id.toString())
+            .subscribe(data => {
+                this.setSuccess(`La organización "${organization.name}" fue eliminada`);
+                this.organizations = this.organizations.filter(org => org.id !== organization.id);
+              },
+              error => this.setError('No se pudo eliminar la organización'));
+        }
+    });
   }
 }
